requirements.txt
requests~=2.32.2
idna~=3.7
certifi~=2024.2.2
urllib3~=2.2.1
Flask~=3.0.3
MarkupSafe~=2.1.5
Werkzeug~=3.0.3
click~=8.1.7
Jinja2~=3.1.4
blinker~=1.8.2
itsdangerous~=2.2.0
python-dotenv~=1.0.1
langchain~=0.2.1
typing_extensions~=4.12.0
attrs~=23.2.0
PyYAML~=6.0.1
multidict~=6.0.5
numpy~=1.26.4
aiohttp~=3.9.5
yarl~=1.9.4
packaging~=23.2
orjson~=3.10.3
aiosignal~=1.3.1
frozenlist~=1.4.1
pydantic~=2.7.1
pydantic_core~=2.18.2
tenacity~=8.3.0
langsmith~=0.1.63
SQLAlchemy~=2.0.30
jsonpointer~=2.4
jsonpatch~=1.33